+++ tests/fixture-factory.spec.ts
@@ -8,13 +8,11 @@ describe('FixtureFactory', () => {
     let existsSyncSpy: jest.SpyInstance;
     let writeFileSyncSpy: jest.SpyInstance;
     let readFileIfExistsSpy: jest.SpyInstance;
-    let accessSyncSpy: jest.SpyInstance;
 
     beforeEach(() => {
         readFileIfExistsSpy = jest.spyOn(fileUtils, 'readFileIfExists');
         readFileIfExistsSpy.mockImplementation(() => defaults);
 
-        accessSyncSpy = jest.spyOn(fs, 'accessSync');
         existsSyncSpy = jest.spyOn(fs, 'existsSync');
         writeFileSyncSpy = jest.spyOn(fs, 'writeFileSync');
         writeFileSyncSpy.mockImplementation(() => undefined);
@@ -27,33 +25,43 @@ describe('FixtureFactory', () => {
     describe('getOrCreateFixture', () => {
         it('throws an error on file write error', () => {
             existsSyncSpy.mockReturnValueOnce(true);
-            accessSyncSpy.mockReturnValueOnce(undefined);
             readFileIfExistsSpy.mockReturnValueOnce(null);
             writeFileSyncSpy.mockImplementationOnce(() => {
                 throw new Error('');
             });
             const factory = new FixtureFactory<ComplexObject>(defaults);
             expect(() => factory.fixtureSync('testfile')).toThrow(
-                ERROR_MESSAGES.FILE_WRITE.replace(':filePath', 'testfile.json'),
+                ERROR_MESSAGES.FILE_WRITE.replace(
+                    ':filePath',
+                    'testfile.json',
+                ).replace(':fileError', ': {}'),
+            );
+        });
+        it('joins file name with factory default path, if provided', () => {
+            readFileIfExistsSpy.mockReturnValueOnce(null);
+            const factory = new FixtureFactory<ComplexObject>(
+                defaults,
+                undefined,
+                '/default/path',
+            );
+            factory.fixtureSync('testfile');
+            expect(writeFileSyncSpy).toHaveBeenCalledWith(
+                '/default/path/testfile.json',
+                expect.any(String),
             );
         });
     });
     describe('.save', () => {
         it('returns old contents if file exists', async () => {
-            existsSyncSpy.mockReturnValueOnce(true);
-            accessSyncSpy.mockReturnValueOnce(undefined);
-
             const factory = new FixtureFactory<ComplexObject>({
                 ...defaults,
                 name: 'differentString',
             });
             const result = await factory.fixture('testfile');
-            expect(existsSyncSpy).toHaveBeenCalled();
             expect(result).toEqual(defaults);
         });
         it('returns new contents if file did not exist', async () => {
             existsSyncSpy.mockReturnValueOnce(true);
-            accessSyncSpy.mockReturnValueOnce(undefined);
             readFileIfExistsSpy.mockReturnValueOnce(null);
 
             const factory = new FixtureFactory<ComplexObject>({
@@ -68,20 +76,15 @@ describe('FixtureFactory', () => {
 
     describe('.saveSync', () => {
         it('returns old contents if file exists', () => {
-            existsSyncSpy.mockReturnValueOnce(true);
-            accessSyncSpy.mockReturnValueOnce(undefined);
-
             const factory = new FixtureFactory<ComplexObject>({
                 ...defaults,
                 name: 'differentString',
             });
             const result = factory.fixtureSync('testfile');
-            expect(existsSyncSpy).toHaveBeenCalled();
             expect(result).toEqual(defaults);
         });
         it('returns new contents if file did not exist', () => {
             existsSyncSpy.mockReturnValueOnce(true);
-            accessSyncSpy.mockReturnValueOnce(undefined);
             readFileIfExistsSpy.mockReturnValueOnce(null);
 
             const factory = new FixtureFactory<ComplexObject>({
@@ -96,8 +99,6 @@ describe('FixtureFactory', () => {
 
     describe('.saveBatch', () => {
         it('returns old contents if file exists', async () => {
-            existsSyncSpy.mockReturnValueOnce(true);
-            accessSyncSpy.mockReturnValueOnce(undefined);
             readFileIfExistsSpy.mockReturnValue([defaults]);
 
             const factory = new FixtureFactory<ComplexObject>({
@@ -105,12 +106,10 @@ describe('FixtureFactory', () => {
                 name: 'differentString',
             });
             const result = await factory.fixtureBatch('testfile', 1);
-            expect(existsSyncSpy).toHaveBeenCalled();
             expect(result).toEqual([defaults]);
         });
         it('returns new contents if file did not exist', async () => {
             existsSyncSpy.mockReturnValueOnce(true);
-            accessSyncSpy.mockReturnValueOnce(undefined);
             readFileIfExistsSpy.mockReturnValueOnce(null);
 
             const factory = new FixtureFactory<ComplexObject>({
@@ -125,8 +124,6 @@ describe('FixtureFactory', () => {
 
     describe('.saveBatchSync', () => {
         it('returns old contents if file exists', () => {
-            existsSyncSpy.mockReturnValueOnce(true);
-            accessSyncSpy.mockReturnValueOnce(undefined);
             readFileIfExistsSpy.mockReturnValueOnce([defaults]);
 
             const factory = new FixtureFactory<ComplexObject>({
@@ -134,12 +131,10 @@ describe('FixtureFactory', () => {
                 name: 'differentString',
             });
             const result = factory.fixtureBatchSync('testfile', 1);
-            expect(existsSyncSpy).toHaveBeenCalled();
             expect(result).toEqual([defaults]);
         });
         it('returns new contents if file did not exist', () => {
             existsSyncSpy.mockReturnValueOnce(true);
-            accessSyncSpy.mockReturnValueOnce(undefined);
             readFileIfExistsSpy.mockReturnValueOnce(null);
 
             const factory = new FixtureFactory<ComplexObject>({